moving generic sympy functionality from derivatives to sympy
